Refactor notification for respnisbility decomposition
