+++ backend/internal/service/notification_service.go
@@ -3,13 +3,7 @@ package service
 import (
 	"context"
 	"errors"
-	"fmt"
-	"net/mail"
-	"net/url"
-	"strings"
 	"time"
-
-	"github.com/google/uuid"
 
 	"github.com/radiation/coyote-ci/backend/internal/domain"
 	"github.com/radiation/coyote-ci/backend/internal/repository"
@@ -49,12 +43,34 @@ const (
 )
 
 type NotificationService struct {
-	repo            repository.NotificationSubscriptionRepository
-	preferencesRepo repository.UserNotificationPreferenceRepository
-	settingsRepo    repository.NotificationInstanceSettingsRepository
-	identityRepo    repository.UserSlackIdentityRepository
-	workspaceRepo   repository.SlackWorkspaceIntegrationRepository
-	now             func() time.Time
+	repo             repository.NotificationSubscriptionRepository
+	targetRepo       notificationTargetRepository
+	subscriptionRepo notificationSubscriptionRepository
+	preferencesRepo  repository.UserNotificationPreferenceRepository
+	settingsRepo     repository.NotificationInstanceSettingsRepository
+	identityRepo     repository.UserSlackIdentityRepository
+	workspaceRepo    repository.SlackWorkspaceIntegrationRepository
+	now              func() time.Time
+}
+
+type notificationTargetRepository interface {
+	CreateTarget(ctx context.Context, target domain.NotificationTarget) (domain.NotificationTarget, error)
+	ListTargets(ctx context.Context) ([]domain.NotificationTarget, error)
+	GetTargetByID(ctx context.Context, id string) (domain.NotificationTarget, error)
+	GetOwnedEmailTargetByUserID(ctx context.Context, userID string) (domain.NotificationTarget, error)
+	SetOwnedEmailTargetEnabled(ctx context.Context, ownerUserID string, enabled bool, updatedAt time.Time) (domain.NotificationTarget, error)
+	EnsureOwnedEmailTargetInitialized(ctx context.Context, input repository.EnsureOwnedNotificationEmailTargetInput) (domain.NotificationTarget, error)
+	UpdateTarget(ctx context.Context, target domain.NotificationTarget) (domain.NotificationTarget, error)
+	DeleteTarget(ctx context.Context, id string) error
+}
+
+type notificationSubscriptionRepository interface {
+	GetTargetByID(ctx context.Context, id string) (domain.NotificationTarget, error)
+	CreateSubscription(ctx context.Context, subscription domain.NotificationSubscription) (domain.NotificationSubscription, error)
+	ListSubscriptions(ctx context.Context, filter repository.NotificationSubscriptionListFilter) ([]domain.NotificationSubscription, error)
+	GetSubscriptionByID(ctx context.Context, id string) (domain.NotificationSubscription, error)
+	UpdateSubscription(ctx context.Context, subscription domain.NotificationSubscription) (domain.NotificationSubscription, error)
+	DeleteSubscription(ctx context.Context, id string) error
 }
 
 type notificationPreferenceRepositoryAware interface {
@@ -66,7 +82,12 @@ type notificationInstanceSettingsRepositoryAware interface {
 }
 
 func NewNotificationService(repo repository.NotificationSubscriptionRepository) *NotificationService {
-	return &NotificationService{repo: repo, now: time.Now}
+	return &NotificationService{
+		repo:             repo,
+		targetRepo:       repo,
+		subscriptionRepo: repo,
+		now:              time.Now,
+	}
 }
 
 type CommitAuthorNotificationPreferenceState struct {
@@ -156,660 +177,4 @@ type CreateNotificationSubscriptionInput struct {
 
 type UpdateNotificationSubscriptionInput struct {
 	Enabled *bool
-}
-
-func (s *NotificationService) ListTargets(ctx context.Context) ([]domain.NotificationTarget, error) {
-	return s.repo.ListTargets(ctx)
-}
-
-func (s *NotificationService) GetOwnedEmailTarget(ctx context.Context, user domain.User) (domain.NotificationTarget, error) {
-	ownerUserID := strings.TrimSpace(user.ID)
-	if ownerUserID == "" {
-		return domain.NotificationTarget{}, ErrNotificationPersonalUserIDRequired
-	}
-	return s.repo.GetOwnedEmailTargetByUserID(ctx, ownerUserID)
-}
-
-func (s *NotificationService) EnsureOwnedEmailTarget(ctx context.Context, user domain.User) (domain.NotificationTarget, error) {
-	ownerUserID := strings.TrimSpace(user.ID)
-	if ownerUserID == "" {
-		return domain.NotificationTarget{}, ErrNotificationPersonalUserIDRequired
-	}
-
-	normalizedIdentityEmail := NormalizeEmail(user.Email)
-	if normalizedIdentityEmail == "" {
-		return domain.NotificationTarget{}, ErrNotificationPersonalEmailRequired
-	}
-
-	recipient, err := normalizeNotificationEmailAddress(normalizedIdentityEmail)
-	if err != nil {
-		if errors.Is(err, ErrNotificationTargetAddressRequired) || errors.Is(err, ErrNotificationTargetAddressInvalid) {
-			return domain.NotificationTarget{}, ErrNotificationPersonalEmailRequired
-		}
-		return domain.NotificationTarget{}, err
-	}
-
-	targetName := normalizedIdentityEmail
-	if user.DisplayName != nil {
-		if trimmedDisplayName := strings.TrimSpace(*user.DisplayName); trimmedDisplayName != "" {
-			targetName = trimmedDisplayName
-		}
-	}
-
-	now := s.now().UTC()
-	input := repository.EnsureOwnedNotificationEmailTargetInput{
-		ID:          uuid.NewString(),
-		OwnerUserID: ownerUserID,
-		Name:        targetName,
-		Recipient:   recipient,
-		CreatedAt:   now,
-		UpdatedAt:   now,
-	}
-	return s.repo.EnsureOwnedEmailTargetInitialized(ctx, input)
-}
-
-func (s *NotificationService) SetOwnedEmailTargetEnabled(ctx context.Context, user domain.User, enabled *bool) (domain.NotificationTarget, error) {
-	ownerUserID := strings.TrimSpace(user.ID)
-	if ownerUserID == "" {
-		return domain.NotificationTarget{}, ErrNotificationPersonalUserIDRequired
-	}
-	if enabled == nil {
-		return domain.NotificationTarget{}, ErrNotificationTargetEnabledRequired
-	}
-
-	return s.repo.SetOwnedEmailTargetEnabled(ctx, ownerUserID, *enabled, s.now().UTC())
-}
-
-func (s *NotificationService) GetNotificationDefaults(ctx context.Context) (NotificationDefaultsState, error) {
-	failureEnabled, successEnabled, err := s.getNotificationDefaultsEnabled(ctx)
-	if err != nil {
-		return NotificationDefaultsState{}, err
-	}
-	return NotificationDefaultsState{
-		DefaultCommitAuthorFailureEmailEnabled: failureEnabled,
-		DefaultCommitAuthorSuccessEmailEnabled: successEnabled,
-	}, nil
-}
-
-func (s *NotificationService) SetNotificationDefaults(ctx context.Context, failureEnabled *bool, successEnabled *bool) (NotificationDefaultsState, error) {
-	if failureEnabled == nil && successEnabled == nil {
-		return NotificationDefaultsState{}, ErrNotificationDefaultsUpdateRequired
-	}
-	if s.settingsRepo == nil {
-		return NotificationDefaultsState{}, errors.New("notification instance settings repository is not configured")
-	}
-	now := s.now().UTC()
-	settings := domain.NotificationInstanceSettings{
-		DefaultCommitAuthorFailureEmailEnabled: true,
-		DefaultCommitAuthorSuccessEmailEnabled: false,
-		CreatedAt:                              now,
-		UpdatedAt:                              now,
-	}
-	createdAt := now
-	if existing, err := s.settingsRepo.Get(ctx); err == nil {
-		settings = existing
-		createdAt = existing.CreatedAt
-	} else if !errors.Is(err, repository.ErrNotificationInstanceSettingsNotFound) {
-		return NotificationDefaultsState{}, err
-	}
-	if failureEnabled != nil {
-		settings.DefaultCommitAuthorFailureEmailEnabled = *failureEnabled
-	}
-	if successEnabled != nil {
-		settings.DefaultCommitAuthorSuccessEmailEnabled = *successEnabled
-	}
-	settings.CreatedAt = createdAt
-	settings.UpdatedAt = now
-	settings, err := s.settingsRepo.Upsert(ctx, settings)
-	if err != nil {
-		return NotificationDefaultsState{}, err
-	}
-	return NotificationDefaultsState{
-		DefaultCommitAuthorFailureEmailEnabled: settings.DefaultCommitAuthorFailureEmailEnabled,
-		DefaultCommitAuthorSuccessEmailEnabled: settings.DefaultCommitAuthorSuccessEmailEnabled,
-	}, nil
-}
-
-func (s *NotificationService) GetCommitAuthorFailureNotificationPreference(ctx context.Context, user domain.User) (CommitAuthorNotificationPreferenceState, error) {
-	ownerUserID := strings.TrimSpace(user.ID)
-	if ownerUserID == "" {
-		return CommitAuthorNotificationPreferenceState{}, ErrNotificationPersonalUserIDRequired
-	}
-
-	preference, err := s.getUserNotificationPreference(ctx, ownerUserID)
-	if err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-	emailState, err := s.resolveEmailPreferenceState(ctx, ownerUserID)
-	if err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-	emailState.Enabled = preference.CommitAuthorFailureEmailEnabled
-	emailState.DeliveryActive = emailState.Enabled && emailState.DeliveryActive
-	slackState, err := s.resolveSlackPreferenceState(ctx, ownerUserID, false)
-	if err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-	slackState.Enabled = preference.CommitAuthorFailureSlackEnabled
-	slackState.DeliveryActive = slackState.Enabled && slackState.DeliveryActive
-	return CommitAuthorNotificationPreferenceState{
-		Email: emailState,
-		Slack: slackState,
-	}, nil
-}
-
-func (s *NotificationService) GetCommitAuthorSuccessNotificationPreference(ctx context.Context, user domain.User) (CommitAuthorNotificationPreferenceState, error) {
-	ownerUserID := strings.TrimSpace(user.ID)
-	if ownerUserID == "" {
-		return CommitAuthorNotificationPreferenceState{}, ErrNotificationPersonalUserIDRequired
-	}
-
-	preference, err := s.getUserNotificationPreference(ctx, ownerUserID)
-	if err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-	emailState, err := s.resolveEmailPreferenceState(ctx, ownerUserID)
-	if err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-	emailState.Enabled = preference.CommitAuthorSuccessEmailEnabled
-	emailState.DeliveryActive = emailState.Enabled && emailState.DeliveryActive
-	slackState, err := s.resolveSlackPreferenceState(ctx, ownerUserID, false)
-	if err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-	slackState.Enabled = preference.CommitAuthorSuccessSlackEnabled
-	slackState.DeliveryActive = slackState.Enabled && slackState.DeliveryActive
-	return CommitAuthorNotificationPreferenceState{
-		Email: emailState,
-		Slack: slackState,
-	}, nil
-}
-
-func (s *NotificationService) SetCommitAuthorFailureNotificationPreference(ctx context.Context, user domain.User, input UpdateCommitAuthorNotificationPreferenceInput) (CommitAuthorNotificationPreferenceState, error) {
-	ownerUserID := strings.TrimSpace(user.ID)
-	if ownerUserID == "" {
-		return CommitAuthorNotificationPreferenceState{}, ErrNotificationPersonalUserIDRequired
-	}
-	if input.EmailEnabled == nil || input.SlackEnabled == nil {
-		return CommitAuthorNotificationPreferenceState{}, ErrNotificationPreferenceChannelEnabledRequired
-	}
-	if err := s.validateCommitAuthorPreferenceChannels(ctx, ownerUserID, *input.EmailEnabled, *input.SlackEnabled); err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-	if s.preferencesRepo == nil {
-		return CommitAuthorNotificationPreferenceState{}, errors.New("notification preference repository is not configured")
-	}
-
-	now := s.now().UTC()
-	preference, err := s.getUserNotificationPreference(ctx, ownerUserID)
-	if err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-	preference.UpdatedAt = now
-	preference.CommitAuthorFailureEmailEnabled = *input.EmailEnabled
-	preference.CommitAuthorFailureSlackEnabled = *input.SlackEnabled
-	preference.CommitAuthorFailureEmailSource = domain.UserNotificationPreferenceSourceUser
-	_, err = s.preferencesRepo.Upsert(ctx, preference)
-	if err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-
-	return s.GetCommitAuthorFailureNotificationPreference(ctx, user)
-}
-
-func (s *NotificationService) SetCommitAuthorSuccessNotificationPreference(ctx context.Context, user domain.User, input UpdateCommitAuthorNotificationPreferenceInput) (CommitAuthorNotificationPreferenceState, error) {
-	ownerUserID := strings.TrimSpace(user.ID)
-	if ownerUserID == "" {
-		return CommitAuthorNotificationPreferenceState{}, ErrNotificationPersonalUserIDRequired
-	}
-	if input.EmailEnabled == nil || input.SlackEnabled == nil {
-		return CommitAuthorNotificationPreferenceState{}, ErrNotificationPreferenceChannelEnabledRequired
-	}
-	if err := s.validateCommitAuthorPreferenceChannels(ctx, ownerUserID, *input.EmailEnabled, *input.SlackEnabled); err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-	if s.preferencesRepo == nil {
-		return CommitAuthorNotificationPreferenceState{}, errors.New("notification preference repository is not configured")
-	}
-
-	now := s.now().UTC()
-	preference, err := s.getUserNotificationPreference(ctx, ownerUserID)
-	if err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-	successSource := domain.UserNotificationPreferenceSourceUser
-	preference.UpdatedAt = now
-	preference.CommitAuthorSuccessEmailEnabled = *input.EmailEnabled
-	preference.CommitAuthorSuccessSlackEnabled = *input.SlackEnabled
-	preference.CommitAuthorSuccessEmailSource = &successSource
-	_, err = s.preferencesRepo.Upsert(ctx, preference)
-	if err != nil {
-		return CommitAuthorNotificationPreferenceState{}, err
-	}
-
-	return s.GetCommitAuthorSuccessNotificationPreference(ctx, user)
-}
-
-func (s *NotificationService) CreateTarget(ctx context.Context, input CreateNotificationTargetInput) (domain.NotificationTarget, error) {
-	name := strings.TrimSpace(input.Name)
-	if name == "" {
-		return domain.NotificationTarget{}, ErrNotificationTargetNameRequired
-	}
-	targetType, err := normalizeNotificationTargetType(input.Type)
-	if err != nil {
-		return domain.NotificationTarget{}, err
-	}
-	recipient, err := normalizeNotificationTargetRecipient(targetType, input.Address, input.WebhookURL)
-	if err != nil {
-		return domain.NotificationTarget{}, err
-	}
-	enabled := true
-	if input.Enabled != nil {
-		enabled = *input.Enabled
-	}
-	now := s.now().UTC()
-
-	return s.repo.CreateTarget(ctx, domain.NotificationTarget{
-		ID:        uuid.NewString(),
-		Type:      targetType,
-		Origin:    domain.NotificationTargetOriginManual,
-		Name:      name,
-		Recipient: recipient,
-		Enabled:   enabled,
-		CreatedAt: now,
-		UpdatedAt: now,
-	})
-}
-
-func (s *NotificationService) CreateEmailTarget(ctx context.Context, input CreateNotificationTargetInput) (domain.NotificationTarget, error) {
-	input.Type = string(domain.NotificationTargetTypeEmail)
-	return s.CreateTarget(ctx, input)
-}
-
-func (s *NotificationService) UpdateTarget(ctx context.Context, id string, input UpdateNotificationTargetInput) (domain.NotificationTarget, error) {
-	targetID, err := normalizeRequiredNotificationUUID(id, repository.ErrNotificationTargetNotFound, ErrNotificationTargetIDInvalid)
-	if err != nil {
-		return domain.NotificationTarget{}, err
-	}
-	current, err := s.repo.GetTargetByID(ctx, targetID)
-	if err != nil {
-		return domain.NotificationTarget{}, err
-	}
-	if input.Name != nil {
-		name := strings.TrimSpace(*input.Name)
-		if name == "" {
-			return domain.NotificationTarget{}, ErrNotificationTargetNameRequired
-		}
-		current.Name = name
-	}
-	if input.Address != nil || input.WebhookURL != nil {
-		address := ""
-		if input.Address != nil {
-			address = *input.Address
-		}
-		webhookURL := ""
-		if input.WebhookURL != nil {
-			webhookURL = *input.WebhookURL
-		}
-		recipient, recipientErr := normalizeNotificationTargetRecipient(current.Type, address, webhookURL)
-		if recipientErr != nil {
-			return domain.NotificationTarget{}, recipientErr
-		}
-		current.Recipient = recipient
-	}
-	if input.Enabled != nil {
-		current.Enabled = *input.Enabled
-	}
-	current.UpdatedAt = s.now().UTC()
-	return s.repo.UpdateTarget(ctx, current)
-}
-
-func (s *NotificationService) DeleteTarget(ctx context.Context, id string) error {
-	targetID, err := normalizeRequiredNotificationUUID(id, repository.ErrNotificationTargetNotFound, ErrNotificationTargetIDInvalid)
-	if err != nil {
-		return err
-	}
-	return s.repo.DeleteTarget(ctx, targetID)
-}
-
-func (s *NotificationService) ListSubscriptions(ctx context.Context, input ListNotificationSubscriptionsInput) ([]domain.NotificationSubscription, error) {
-	projectID, err := normalizeOptionalNotificationUUID(input.ProjectID, ErrNotificationSubscriptionProjectIDInvalid)
-	if err != nil {
-		return nil, err
-	}
-	jobID, err := normalizeOptionalNotificationUUID(input.JobID, ErrNotificationSubscriptionJobIDInvalid)
-	if err != nil {
-		return nil, err
-	}
-	return s.repo.ListSubscriptions(ctx, repository.NotificationSubscriptionListFilter{
-		ProjectID: projectID,
-		JobID:     jobID,
-	})
-}
-
-func (s *NotificationService) CreateSubscription(ctx context.Context, input CreateNotificationSubscriptionInput) (domain.NotificationSubscription, error) {
-	targetID, err := normalizeRequiredNotificationUUID(input.TargetID, ErrNotificationSubscriptionTargetIDRequired, ErrNotificationSubscriptionTargetIDInvalid)
-	if err != nil {
-		return domain.NotificationSubscription{}, err
-	}
-	if _, getTargetErr := s.repo.GetTargetByID(ctx, targetID); getTargetErr != nil {
-		return domain.NotificationSubscription{}, getTargetErr
-	}
-	projectID, err := normalizeOptionalNotificationUUID(input.ProjectID, ErrNotificationSubscriptionProjectIDInvalid)
-	if err != nil {
-		return domain.NotificationSubscription{}, err
-	}
-	jobID, err := normalizeOptionalNotificationUUID(input.JobID, ErrNotificationSubscriptionJobIDInvalid)
-	if err != nil {
-		return domain.NotificationSubscription{}, err
-	}
-	if (projectID == nil) == (jobID == nil) {
-		return domain.NotificationSubscription{}, ErrNotificationSubscriptionScopeRequired
-	}
-	eventType, err := normalizeNotificationEventType(input.EventType)
-	if err != nil {
-		return domain.NotificationSubscription{}, err
-	}
-	enabled := true
-	if input.Enabled != nil {
-		enabled = *input.Enabled
-	}
-	now := s.now().UTC()
-
-	return s.repo.CreateSubscription(ctx, domain.NotificationSubscription{
-		ID:        uuid.NewString(),
-		TargetID:  targetID,
-		ProjectID: projectID,
-		JobID:     jobID,
-		EventType: eventType,
-		Enabled:   enabled,
-		CreatedAt: now,
-		UpdatedAt: now,
-	})
-}
-
-func (s *NotificationService) UpdateSubscription(ctx context.Context, id string, input UpdateNotificationSubscriptionInput) (domain.NotificationSubscription, error) {
-	subscriptionID, err := normalizeRequiredNotificationUUID(id, repository.ErrNotificationSubscriptionNotFound, ErrNotificationSubscriptionIDInvalid)
-	if err != nil {
-		return domain.NotificationSubscription{}, err
-	}
-	current, err := s.repo.GetSubscriptionByID(ctx, subscriptionID)
-	if err != nil {
-		return domain.NotificationSubscription{}, err
-	}
-	if input.Enabled != nil {
-		current.Enabled = *input.Enabled
-	}
-	current.UpdatedAt = s.now().UTC()
-	return s.repo.UpdateSubscription(ctx, current)
-}
-
-func (s *NotificationService) DeleteSubscription(ctx context.Context, id string) error {
-	subscriptionID, err := normalizeRequiredNotificationUUID(id, repository.ErrNotificationSubscriptionNotFound, ErrNotificationSubscriptionIDInvalid)
-	if err != nil {
-		return err
-	}
-	return s.repo.DeleteSubscription(ctx, subscriptionID)
-}
-
-func normalizeNotificationEmailAddress(value string) (string, error) {
-	trimmed := strings.TrimSpace(value)
-	if trimmed == "" {
-		return "", ErrNotificationTargetAddressRequired
-	}
-	address, err := mail.ParseAddress(trimmed)
-	if err != nil {
-		return "", ErrNotificationTargetAddressInvalid
-	}
-	return (&mail.Address{Address: NormalizeEmail(address.Address)}).String(), nil
-}
-
-func normalizeNotificationTargetType(value string) (domain.NotificationTargetType, error) {
-	trimmed := strings.TrimSpace(value)
-	if trimmed == "" {
-		return domain.NotificationTargetTypeEmail, nil
-	}
-	switch domain.NotificationTargetType(trimmed) {
-	case domain.NotificationTargetTypeEmail, domain.NotificationTargetTypeSlackWebhook:
-		return domain.NotificationTargetType(trimmed), nil
-	default:
-		return "", ErrNotificationTargetTypeInvalid
-	}
-}
-
-func normalizeNotificationTargetRecipient(targetType domain.NotificationTargetType, address string, webhookURL string) (string, error) {
-	switch targetType {
-	case domain.NotificationTargetTypeEmail:
-		return normalizeNotificationEmailAddress(address)
-	case domain.NotificationTargetTypeSlackWebhook:
-		return normalizeNotificationWebhookURL(webhookURL)
-	default:
-		return "", ErrNotificationTargetTypeInvalid
-	}
-}
-
-func (s *NotificationService) getCommitAuthorFailurePreferenceEnabled(ctx context.Context, userID string) (bool, error) {
-	if s.preferencesRepo == nil {
-		return false, nil
-	}
-	preference, err := s.preferencesRepo.GetByUserID(ctx, userID)
-	if err != nil {
-		if errors.Is(err, repository.ErrUserNotificationPreferenceNotFound) {
-			return false, nil
-		}
-		return false, err
-	}
-	return preference.CommitAuthorFailureEmailEnabled, nil
-}
-
-func (s *NotificationService) getCommitAuthorSuccessPreferenceEnabled(ctx context.Context, userID string) (bool, error) {
-	if s.preferencesRepo == nil {
-		return false, nil
-	}
-	preference, err := s.preferencesRepo.GetByUserID(ctx, userID)
-	if err != nil {
-		if errors.Is(err, repository.ErrUserNotificationPreferenceNotFound) {
-			return false, nil
-		}
-		return false, err
-	}
-	return preference.CommitAuthorSuccessEmailEnabled, nil
-}
-
-func (s *NotificationService) getUserNotificationPreference(ctx context.Context, userID string) (domain.UserNotificationPreference, error) {
-	if s.preferencesRepo == nil {
-		return domain.UserNotificationPreference{UserID: userID}, nil
-	}
-	now := s.now().UTC()
-	preference := domain.UserNotificationPreference{
-		UserID:                         userID,
-		CommitAuthorFailureEmailSource: domain.UserNotificationPreferenceSourceUser,
-		CreatedAt:                      now,
-		UpdatedAt:                      now,
-	}
-	stored, err := s.preferencesRepo.GetByUserID(ctx, userID)
-	if err != nil {
-		if errors.Is(err, repository.ErrUserNotificationPreferenceNotFound) {
-			return preference, nil
-		}
-		return domain.UserNotificationPreference{}, err
-	}
-	return stored, nil
-}
-
-func (s *NotificationService) resolveEmailPreferenceState(ctx context.Context, userID string) (CommitAuthorEmailNotificationPreferenceState, error) {
-	state := CommitAuthorEmailNotificationPreferenceState{}
-	target, err := s.repo.GetOwnedEmailTargetByUserID(ctx, userID)
-	if err != nil {
-		if errors.Is(err, repository.ErrNotificationTargetNotFound) {
-			reason := NotificationPreferenceUnavailableReasonPersonalTargetRequired
-			state.UnavailableReason = &reason
-			return state, nil
-		}
-		return CommitAuthorEmailNotificationPreferenceState{}, err
-	}
-	state.Target = &target
-	state.DeliveryActive = target.Enabled
-	if !target.Enabled {
-		reason := NotificationPreferenceUnavailableReasonPersonalTargetDisabled
-		state.UnavailableReason = &reason
-	}
-	return state, nil
-}
-
-func (s *NotificationService) resolveSlackPreferenceState(ctx context.Context, userID string, requireDeliverable bool) (CommitAuthorSlackNotificationPreferenceState, error) {
-	state := CommitAuthorSlackNotificationPreferenceState{}
-	if s.workspaceRepo == nil || s.identityRepo == nil {
-		reason := NotificationPreferenceUnavailableReasonSlackWorkspaceNotConfigured
-		state.UnavailableReason = &reason
-		return state, nil
-	}
-	integration, err := s.workspaceRepo.Get(ctx)
-	if err != nil {
-		if errors.Is(err, repository.ErrSlackWorkspaceIntegrationNotFound) {
-			reason := NotificationPreferenceUnavailableReasonSlackWorkspaceNotConfigured
-			state.UnavailableReason = &reason
-			return state, nil
-		}
-		return CommitAuthorSlackNotificationPreferenceState{}, err
-	}
-	if strings.TrimSpace(integration.BotTokenSecret) == "" {
-		reason := NotificationPreferenceUnavailableReasonSlackWorkspaceNotConfigured
-		state.UnavailableReason = &reason
-		return state, nil
-	}
-	if !integration.Enabled {
-		reason := NotificationPreferenceUnavailableReasonSlackWorkspaceDisabled
-		state.UnavailableReason = &reason
-		return state, nil
-	}
-	identity, err := s.identityRepo.GetByUserID(ctx, userID)
-	if err != nil {
-		if errors.Is(err, repository.ErrUserSlackIdentityNotFound) {
-			reason := NotificationPreferenceUnavailableReasonSlackIdentityRequired
-			state.UnavailableReason = &reason
-			return state, nil
-		}
-		return CommitAuthorSlackNotificationPreferenceState{}, err
-	}
-	if strings.TrimSpace(identity.SlackUserID) == "" {
-		reason := NotificationPreferenceUnavailableReasonSlackIdentityRequired
-		state.UnavailableReason = &reason
-		return state, nil
-	}
-	if !identity.Enabled {
-		reason := NotificationPreferenceUnavailableReasonSlackIdentityDisabled
-		state.UnavailableReason = &reason
-		return state, nil
-	}
-	if identity.SlackWorkspaceIntegrationID != integration.ID {
-		reason := NotificationPreferenceUnavailableReasonSlackWorkspaceMismatch
-		state.UnavailableReason = &reason
-		return state, nil
-	}
-	state.DeliveryActive = true
-	if requireDeliverable {
-		return state, nil
-	}
-	return state, nil
-}
-
-func (s *NotificationService) validateCommitAuthorPreferenceChannels(ctx context.Context, userID string, emailEnabled bool, slackEnabled bool) error {
-	if emailEnabled {
-		emailState, err := s.resolveEmailPreferenceState(ctx, userID)
-		if err != nil {
-			return err
-		}
-		if !emailState.DeliveryActive {
-			return ErrNotificationPreferencePersonalTargetRequired
-		}
-	}
-	if slackEnabled {
-		state, err := s.resolveSlackPreferenceState(ctx, userID, true)
-		if err != nil {
-			return err
-		}
-		if !state.DeliveryActive {
-			return ErrNotificationPreferencePersonalSlackRequired
-		}
-	}
-	return nil
-}
-
-func (s *NotificationService) getDefaultCommitAuthorFailureEmailEnabled(ctx context.Context) (bool, error) {
-	enabled, _, err := s.getNotificationDefaultsEnabled(ctx)
-	return enabled, err
-}
-
-func (s *NotificationService) getNotificationDefaultsEnabled(ctx context.Context) (bool, bool, error) {
-	if s.settingsRepo == nil {
-		return true, false, nil
-	}
-	settings, err := s.settingsRepo.Get(ctx)
-	if err != nil {
-		if errors.Is(err, repository.ErrNotificationInstanceSettingsNotFound) {
-			return true, false, nil
-		}
-		return false, false, fmt.Errorf("get notification instance settings: %w", err)
-	}
-	return settings.DefaultCommitAuthorFailureEmailEnabled, settings.DefaultCommitAuthorSuccessEmailEnabled, nil
-}
-
-func normalizeNotificationWebhookURL(value string) (string, error) {
-	trimmed := strings.TrimSpace(value)
-	if trimmed == "" {
-		return "", ErrNotificationTargetWebhookURLRequired
-	}
-	parsed, err := url.Parse(trimmed)
-	if err != nil || parsed == nil || !parsed.IsAbs() || strings.ToLower(parsed.Scheme) != "https" || strings.TrimSpace(parsed.Host) == "" {
-		return "", ErrNotificationTargetWebhookURLInvalid
-	}
-	return parsed.String(), nil
-}
-
-func normalizeNotificationEventType(value string) (domain.NotificationEventType, error) {
-	trimmed := strings.TrimSpace(value)
-	switch domain.NotificationEventType(trimmed) {
-	case domain.NotificationEventTypeBuildSucceeded, domain.NotificationEventTypeBuildFailed:
-		return domain.NotificationEventType(trimmed), nil
-	default:
-		return "", ErrNotificationSubscriptionEventTypeInvalid
-	}
-}
-
-func trimOptionalStringValue(value *string) *string {
-	if value == nil {
-		return nil
-	}
-	trimmed := strings.TrimSpace(*value)
-	if trimmed == "" {
-		return nil
-	}
-	return &trimmed
-}
-
-func normalizeRequiredNotificationUUID(value string, emptyErr error, invalidErr error) (string, error) {
-	trimmed := strings.TrimSpace(value)
-	if trimmed == "" {
-		return "", emptyErr
-	}
-	parsed, parseErr := uuid.Parse(trimmed)
-	if parseErr != nil {
-		return "", invalidErr
-	}
-	return parsed.String(), nil
-}
-
-func normalizeOptionalNotificationUUID(value *string, invalidErr error) (*string, error) {
-	trimmed := trimOptionalStringValue(value)
-	if trimmed == nil {
-		return nil, nil
-	}
-	parsed, parseErr := uuid.Parse(*trimmed)
-	if parseErr != nil {
-		return nil, invalidErr
-	}
-	normalized := parsed.String()
-	return &normalized, nil
 }
